add tests for converter(&simulator)
