+++ test/with_qiskit/simulator/state_vector_circuit/test_rx_gate.py
@@ -4,8 +4,10 @@ import numpy as np
 from qiskit import QuantumCircuit
 from qiskit.quantum_info.operators import Operator
 
-from quantestpy import StateVectorCircuit
-from quantestpy.simulator.state_vector_circuit import _rx
+from quantestpy.converter.converter_to_quantestpy_circuit import \
+    cvt_input_circuit_to_quantestpy_circuit
+from quantestpy.simulator.state_vector_circuit import \
+    cvt_quantestpy_circuit_to_state_vector_circuit
 
 
 class TestStateVectorCircuitRxGate(unittest.TestCase):
@@ -13,220 +15,64 @@ class TestStateVectorCircuitRxGate(unittest.TestCase):
     How to execute this test:
     $ pwd
     {Your directory where you git-cloned quantestpy}/quantestpy
-    $ python -m unittest test.simulator.state_vector_circuit.test_rx_gate
-    ...........
+    $ python -m unittest \
+        test.with_qiskit.simulator.state_vector_circuit.test_rx_gate
+    ....
     ----------------------------------------------------------------------
-    Ran 11 tests in 0.012s
+    Ran 4 tests in 0.010s
 
     OK
     $
     """
 
-    def test_crx_regular_qubit_order(self,):
-        circ = StateVectorCircuit(2)
-        lambda_ = np.pi/8
-        actual_gate = circ._create_all_qubit_gate_from_original_qubit_gate(
-            _rx([lambda_]),
-            control_qubit=[0], target_qubit=[1], control_value=[1])
-
-        expected_gate = np.array([
-            [1, 0, 0, 0],
-            [0, 1, 0, 0],
-            [0, 0, np.cos(lambda_/2), -1j*np.sin(lambda_/2)],
-            [0, 0, -1j*np.sin(lambda_/2), np.cos(lambda_/2)]])
-
-        self.assertIsNone(
-            np.testing.assert_allclose(actual_gate, expected_gate))
-
-    def test_crx_qiskit_qubit_order(self,):
-        circ = StateVectorCircuit(2)
-        lambda_ = np.pi/8
-        circ._from_right_to_left_for_qubit_ids = True
-        actual_gate = circ._create_all_qubit_gate_from_original_qubit_gate(
-            _rx([lambda_]),
-            control_qubit=[0], target_qubit=[1], control_value=[1])
-
-        expected_gate = np.array([
-            [1, 0, 0, 0],
-            [0, np.cos(lambda_/2), 0, -1j*np.sin(lambda_/2)],
-            [0, 0, 1, 0],
-            [0, - 1j*np.sin(lambda_/2), 0, np.cos(lambda_/2)]])
-
-        self.assertIsNone(
-            np.testing.assert_allclose(actual_gate, expected_gate))
-
-    def test_crx_flip_control_target(self,):
-        circ = StateVectorCircuit(2)
-        lambda_ = np.pi/8
-        actual_gate = circ._create_all_qubit_gate_from_original_qubit_gate(
-            _rx([lambda_]),
-            control_qubit=[1], target_qubit=[0], control_value=[1])
-
-        expected_gate = np.array([
-            [1, 0, 0, 0],
-            [0, np.cos(lambda_/2), 0, -1j*np.sin(lambda_/2)],
-            [0, 0, 1, 0],
-            [0, - 1j*np.sin(lambda_/2), 0, np.cos(lambda_/2)]])
-
-        self.assertIsNone(
-            np.testing.assert_allclose(actual_gate, expected_gate))
-
-    def test_crx_three_qubits_qiskit_qubit_order(self,):
-        circ = StateVectorCircuit(3)
-        lambda_ = np.pi/8
-        circ._from_right_to_left_for_qubit_ids = True
-        actual_gate = circ._create_all_qubit_gate_from_original_qubit_gate(
-            _rx([lambda_]),
-            control_qubit=[0], target_qubit=[2], control_value=[1])
-
-        # this is qiskit's output
-        expected_gate = np.array(
-            [[1. + 0.j, 0. + 0.j,
-              0. + 0.j, 0. + 0.j,
-              0. + 0.j, 0. + 0.j,
-              0. + 0.j, 0. + 0.j],
-             [0. + 0.j, 0.98078528+0.j,
-                0. + 0.j, 0. + 0.j,
-                0. + 0.j, 0. - 0.19509032j,
-              0. + 0.j, 0. + 0.j],
-             [0. + 0.j, 0. + 0.j,
-                1. + 0.j, 0. + 0.j,
-                0. + 0.j, 0. + 0.j,
-                0. + 0.j, 0. + 0.j],
-                [0. + 0.j, 0. + 0.j,
-                 0. + 0.j, 0.98078528+0.j,
-                 0. + 0.j, 0. + 0.j,
-                 0. + 0.j, 0. - 0.19509032j],
-                [0. + 0.j, 0. + 0.j,
-                 0. + 0.j, 0. + 0.j,
-                 1. + 0.j, 0. + 0.j,
-                 0. + 0.j, 0. + 0.j],
-                [0. + 0.j, 0. - 0.19509032j,
-                 0. + 0.j, 0. + 0.j,
-                 0. + 0.j, 0.98078528+0.j,
-                 0. + 0.j, 0. + 0.j],
-                [0. + 0.j, 0. + 0.j,
-                 0. + 0.j, 0. + 0.j,
-                 0. + 0.j, 0. + 0.j,
-                 1. + 0.j, 0. + 0.j],
-                [0. + 0.j, 0. + 0.j,
-                 0. + 0.j, 0. - 0.19509032j,
-                 0. + 0.j, 0. + 0.j,
-                 0. + 0.j, 0.98078528+0.j]])
-
-        self.assertIsNone(
-            np.testing.assert_allclose(actual_gate, expected_gate))
-
-    def test_crx_control_value_is_zero(self,):
-        circ = StateVectorCircuit(2)
-        lambda_ = np.pi/8
-        actual_gate = circ._create_all_qubit_gate_from_original_qubit_gate(
-            _rx([lambda_]),
-            control_qubit=[0], target_qubit=[1], control_value=[0])
-
-        expected_gate = np.array([
-            [np.cos(lambda_/2), -1j*np.sin(lambda_/2), 0, 0],
-            [-1j*np.sin(lambda_/2), np.cos(lambda_/2), 0, 0],
-            [0, 0, 1, 0],
-            [0, 0, 0, 1]])
-
-        self.assertIsNone(
-            np.testing.assert_allclose(actual_gate, expected_gate))
-
-    def test_crx_multiple_controls(self,):
-        circ = StateVectorCircuit(3)
-        lambda_ = np.pi/8
-        actual_gate = circ._create_all_qubit_gate_from_original_qubit_gate(
-            _rx([lambda_]),
-            control_qubit=[0, 1], target_qubit=[2], control_value=[1, 1])
-
-        expected_gate = np.array([[1, 0, 0, 0, 0, 0, 0, 0],
-                                  [0, 1, 0, 0, 0, 0, 0, 0],
-                                  [0, 0, 1, 0, 0, 0, 0, 0],
-                                  [0, 0, 0, 1, 0, 0, 0, 0],
-                                  [0, 0, 0, 0, 1, 0, 0, 0],
-                                  [0, 0, 0, 0, 0, 1, 0, 0],
-                                  [0, 0, 0, 0, 0, 0,
-                                   np.cos(lambda_/2), -1j*np.sin(lambda_/2)],
-                                  [0, 0, 0, 0, 0, 0,
-                                   -1j*np.sin(lambda_/2), np.cos(lambda_/2)]])
-
-        self.assertIsNone(
-            np.testing.assert_allclose(actual_gate, expected_gate))
-
-    def test_crx_multiple_targets(self,):
-        lambda_ = np.pi/8
-
-        circ = StateVectorCircuit(3)
-        gate_0 = circ._create_all_qubit_gate_from_original_qubit_gate(
-            _rx([lambda_]),
-            control_qubit=[0], target_qubit=[1, 2], control_value=[1])
-
-        circ = StateVectorCircuit(3)
-        gate_1_0 = circ._create_all_qubit_gate_from_original_qubit_gate(
-            _rx([lambda_]),
-            control_qubit=[0], target_qubit=[1], control_value=[1])
-        gate_1_1 = circ._create_all_qubit_gate_from_original_qubit_gate(
-            _rx([lambda_]),
-            control_qubit=[0], target_qubit=[2], control_value=[1])
-
-        self.assertIsNone(
-            np.testing.assert_allclose(gate_0, np.matmul(gate_1_0, gate_1_1)))
-
     def test_crx_control_value_1(self,):
-        circ = StateVectorCircuit(3)
-        lambda_ = np.pi/4
-        actual_gate = circ._create_all_qubit_gate_from_original_qubit_gate(
-            _rx([lambda_]),
-            control_qubit=[2], target_qubit=[1], control_value=[1])
-
         qc = QuantumCircuit(3)
         qc.crx(np.pi/4, 0, 1)
         expected_gate = np.array(Operator(qc))
+
+        qpc = cvt_input_circuit_to_quantestpy_circuit(qc)
+        svc = cvt_quantestpy_circuit_to_state_vector_circuit(qpc)
+        svc._from_right_to_left_for_qubit_ids = True
+        actual_gate = svc._get_whole_gates()
 
         self.assertIsNone(
             np.testing.assert_allclose(actual_gate, expected_gate))
 
     def test_crx_control_value_0(self,):
-        circ = StateVectorCircuit(3)
-        lambda_ = np.pi/4
-        actual_gate = circ._create_all_qubit_gate_from_original_qubit_gate(
-            _rx([lambda_]),
-            control_qubit=[2], target_qubit=[1], control_value=[0])
-
         qc = QuantumCircuit(3)
         qc.crx(np.pi/4, 0, 1, None, "0")
         expected_gate = np.array(Operator(qc))
+
+        qpc = cvt_input_circuit_to_quantestpy_circuit(qc)
+        svc = cvt_quantestpy_circuit_to_state_vector_circuit(qpc)
+        svc._from_right_to_left_for_qubit_ids = True
+        actual_gate = svc._get_whole_gates()
 
         self.assertIsNone(
             np.testing.assert_allclose(actual_gate, expected_gate))
 
     def test_rx(self,):
-        parameter = [np.pi/4]
-        circ = StateVectorCircuit(3)
-        actual_gate = circ._create_all_qubit_gate_from_original_qubit_gate(
-            _rx(parameter),
-            control_qubit=[], target_qubit=[2], control_value=[]
-        )
-
         qc = QuantumCircuit(3)
         qc.rx(np.pi/4, 0)
         expected_gate = np.array(Operator(qc))
+
+        qpc = cvt_input_circuit_to_quantestpy_circuit(qc)
+        svc = cvt_quantestpy_circuit_to_state_vector_circuit(qpc)
+        svc._from_right_to_left_for_qubit_ids = True
+        actual_gate = svc._get_whole_gates()
 
         self.assertIsNone(
             np.testing.assert_allclose(actual_gate, expected_gate))
 
     def test_mcrx(self,):
-        parameter = [np.pi/4]
-        circ = StateVectorCircuit(3)
-        actual_gate = circ._create_all_qubit_gate_from_original_qubit_gate(
-            _rx(parameter),
-            control_qubit=[2, 1], target_qubit=[0], control_value=[1, 1]
-        )
-
         qc = QuantumCircuit(3)
         qc.mcrx(np.pi/4, [0, 1], 2)
         expected_gate = np.array(Operator(qc))
+
+        qpc = cvt_input_circuit_to_quantestpy_circuit(qc)
+        svc = cvt_quantestpy_circuit_to_state_vector_circuit(qpc)
+        svc._from_right_to_left_for_qubit_ids = True
+        actual_gate = svc._get_whole_gates()
 
         self.assertIsNone(
             np.testing.assert_allclose(actual_gate, expected_gate))
