+++ test/with_qiskit/simulator/state_vector_circuit/test_sxdg_gate.py
@@ -4,8 +4,10 @@ import numpy as np
 from qiskit import QuantumCircuit
 from qiskit.quantum_info.operators import Operator
 
-from quantestpy import StateVectorCircuit
-from quantestpy.simulator.state_vector_circuit import _SXdg
+from quantestpy.converter.converter_to_quantestpy_circuit import \
+    cvt_input_circuit_to_quantestpy_circuit
+from quantestpy.simulator.state_vector_circuit import \
+    cvt_quantestpy_circuit_to_state_vector_circuit
 
 
 class TestStateVectorCircuitSXdgGate(unittest.TestCase):
@@ -13,7 +15,8 @@ class TestStateVectorCircuitSXdgGate(unittest.TestCase):
     How to execute this test:
     $ pwd
     {Your directory where you git-cloned quantestpy}/quantestpy
-    $ python -m unittest test.simulator.state_vector_circuit.test_sxdg_gate
+    $ python -m unittest \
+        test.with_qiskit.simulator.state_vector_circuit.test_sxdg_gate
     .
     ----------------------------------------------------------------------
     Ran 1 test in 0.004s
@@ -23,14 +26,14 @@ class TestStateVectorCircuitSXdgGate(unittest.TestCase):
     """
 
     def test_sxdg(self,):
-        circ = StateVectorCircuit(3)
-        actual_gate = circ._create_all_qubit_gate_from_original_qubit_gate(
-            _SXdg, control_qubit=[], target_qubit=[2], control_value=[]
-        )
-
         qc = QuantumCircuit(3)
         qc.sxdg(0)
         expected_gate = np.array(Operator(qc))
+
+        qpc = cvt_input_circuit_to_quantestpy_circuit(qc)
+        svc = cvt_quantestpy_circuit_to_state_vector_circuit(qpc)
+        svc._from_right_to_left_for_qubit_ids = True
+        actual_gate = svc._get_whole_gates()
 
         self.assertIsNone(
             np.testing.assert_allclose(actual_gate, expected_gate))
